Fixed vector not clearing leading to crash

+++ scripts/autoloads/GDNShell.gd
@@ -4,45 +4,64 @@ const APP_NAME = "GDNShell"
 var LOG = []
 
 const GDNSHELL_RES = preload("res://gdnative.gdns")
-var GDNSHELL = null
+onready var GDNSHELL = GDNSHELL_RES.new()
 
 var root_node = null
 var terminal_node = null
-var shell_thread = null
+onready var shell_thread = Thread.new()
 
+func _child_process_started():
+	print("STARTED!")
+	pass
+func _child_process_stopped():
+	print("STOPPED!")
+	pass
+
+func GDN_INIT():
+	assert(GDNSHELL != null)
+	Log.generic(self, "Loading GDNative...")
+	GDNSHELL.connect("child_process_started", self, "_child_process_started")
+	GDNSHELL.connect("child_process_stopped", self, "_child_process_stopped")
+
+var is_stopping = false
 func thread_subroutine(cmd):
-	Log.generic(self, "Spawning child process...")
-	GDNSHELL = GDNSHELL_RES.new()
-	GDNSHELL.connect("child_process_started", terminal_node, "_child_process_started")
-	GDNSHELL.connect("waiting_for_inputs", terminal_node, "_waiting_for_inputs")
-	GDNSHELL.spawn(terminal_node, cmd) # this will internally loop and update, function will resume on close
-	GDNSHELL = null
-	terminal_node._child_process_stopped()
+	# this will internally loop and update, function will resume on close
+	Log.generic(self, str(":: Entering thread ", shell_thread.get_id()))
+	GDNSHELL.spawn(cmd)
+
+	# cleanup thread
+	Log.generic(self, str(":: Exiting thread ", shell_thread.get_id()))
 
 var shell_cmd = "E:/Git/CppTestApp/cmake-build-debug/CppTestApp.exe"
 func start():
+	assert(GDNSHELL != null)
+
 	# stop any running session first
-	if GDNSHELL != null:
-		stop()
+	stop()
 
 	# clear terminal & start thread
 	clear()
-	shell_thread = Thread.new()
 	shell_thread.start(self, "thread_subroutine", shell_cmd)
 func stop():
-	if GDNSHELL != null:
-		Log.generic(self, "Killing child process...")
+	assert(GDNSHELL != null)
+	if is_stopping:
+		return
+	is_stopping = true
+	if GDNSHELL != null && shell_thread != null && shell_thread.is_active():
+		Log.generic(self, str("Killing child process on thread ", shell_thread.get_id()))
 		GDNSHELL.kill()
-		shell_thread.wait_to_finish()
+		if shell_thread.is_alive():
+			shell_thread.wait_to_finish()
 	else:
 		Log.generic(self, "No process is running!")
+	is_stopping = false
 
 
-func sync_text():
+func sync_text(line, size):
 	if GDNSHELL != null:
-		var buffer = GDNSHELL.fetch()
-		if buffer != "":
-			terminal_node._receive_text(buffer)
+		var buffer = GDNSHELL.fetch_at_line(line, size)
+		return buffer
+	return null
 
 func send_string(s : String):
 	if shell_thread != null:
@@ -51,7 +70,11 @@ func send_string(s : String):
 		else:
 			print("failure...")
 func clear():
+	return
+	assert(terminal_node != null)
+	assert(GDNSHELL != null)
 	terminal_node.clear()
+	GDNSHELL.clear()
 
 func _exit_tree():
 	# close any open session when shutting down
